+++ previous-versions/accounting/RELEASES.md
@@ -391,3 +391,13 @@ Based on:
 - [go v0.27.1] previous-versions/accounting
 ### Releases
 - [Go v0.27.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.27.1 - previous-versions/accounting
+
+## 2023-09-29 00:21:20
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.28.0] previous-versions/accounting
+### Releases
+- [Go v0.28.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/accounting/v0.28.0 - previous-versions/accounting